I need to train some shape predictor and I found that this process is very slow for me - up to 5-6 days with 64 GB RAM usage. Also training process uses a lot of RAM

With this PR I made this changes:
1. training_sample::feature_pixel_values was float-based. Now I am using the type provided by image_traits. In the most cases - it is unsigned char leading to lower RAM usage (about 15% less summary with all other changes). Train_shape_predictor_ex uses 400 features inside each sample (1600 bytes was, now - 400)
2. some parts of trainer uses the shape difference training_sample::target_shape - current shape. And this differenct is recalculated for the same sample many times. I made caching for it inside training_sample::diff_shape. This increases RAM usage a little (+68*2*4 = 544), and improves performance for about 20%
3. After caching differences I found that some parts are now easy to run parallel and I made it work with parallel_for. When I am running it with 8 threads on i7, about 70% of CPU is used (because not all parts are paralleled) and training time reduces 2-3x

It is also possible to remove caching the difference of shapes and use much less RAM (about 20-30% less) depending on the task. 

I have tested this with train_shape_predictor_ex with increasing oversampling_amount to 3000, but I think we need some more time for testing

What do you think? Will such kind of optimizations be usefull?