@davisking 2016-06-03
Sweet.  I looked it over and it seems reasonable to me.  I didn't test it though as I'm away on vacation this week.  I'll look at it more when I get back.  But if you want to test it you should train a landmarking model using this data: http://dlib.net/files/data/ibug_300W_large_face_landmark_dataset.tar.gz.  If you make a model with the default training parameters both before and after this change and get the same results then everything is good.