@davisking 2016-06-06
I ran a bunch of tests on this today.  On a quad core i7, if I use 4 cores it's about 33% faster to train.  But for 1 core it's much slower, taking twice the time it normally does.  For 2 cores I got a slight slowdown but only a little.   Also, in each case the landmarking accuracy is now slightly worse, 0.5%.  It's small and might just be chance.   It doesn't seem like the code changes should result in any real changes to the behavior. I ran a few different tests and it seemed like it's persistently that much worse but it's hard to tell.  Also, the memory usage wasn't substantially different between the two versions.

So here is what I suggest.  Can you make it so using 1 thread doesn't have a slowdown relative to the previous version?  That could also use the lower RAM version you mentioned.  Also, can you make it compile without C++11 support?  I don't want to force C++11 usage yet for dlib users.  That does make the parallel_for code pretty ugly though.  

I'll probably require C++11 sometime around January 2017, so you could also wait until then to merge the PR if you want.  I know there are a bunch of corporate users of dlib who are stuck on old compilers and it takes them about 6 years to update :(.  Even 2017 is a little aggressive for some of them but not using C++11 in the bulk of dlib's code is really starting to grate on me.  So 2017 is the date I'm officially switching dlib to C++11.