@e-fominov 2016-06-07
Thanks for the tesing. My servers are still busy. Planning to re-check it on this week
I will find some good way to solve this problems