@davisking 2016-06-07
Cool.

I should also mention that the main holdup on C++11 is visual studio.  If this code works in visual studio 2013 maybe we can include it.  Do you know what version range of visual studio supports these features?