@e-fominov 2016-06-10
Fixed compatibility for Visual Studio. Tested VS2010, VS2015, MINGW 5.3. 
Fixed performance leak in single thread mode (reason of problem - optimizer and CPU cache). Now trainer.set_num_threads(0 or 1) will make it work no slower than original implementation (really 20% faster)

I am not finished checks with ibug - waiting for complete.
0 and 1 thread mode now have no difference in the testing scores with original version
2+ thread mode has some difference in the 4th digit normally. This is because rounding and limited float 32 precision. I can make it work with no difference, but performance affects very much

My measurements on the examples\train_shape_predictor_ex (VS2015, x64)

|Threads|Time, seconds|Training error|Testing error|
----------|------------------|-----------------|----------------
|original | 171 sec|2.40105e-05|0.085379|
|new, 0-1|148 sec|2.40105e-05|0.085379|
|2          |96 sec|0 |0.0856668|
|3          |70 sec|2.40105e-05|0.0856295 |
|8          |59 sec|2.40105e-05 | 0.0858176|

difference between 3 and 8 threads is small, because sample uses small amount of data. testing results on ibug dataset - tomorrow
in 8 thread mode, CPU utilization on i7 is about 60-70%, because some parts are hard to make parallel