@davisking 2016-06-11
Sweet, nice work.  Let me know when it's ready.

I've also been thinking about C++11 and think waiting until 2017 is
excessive.  I'll make the next release of dlib soon, which has some C++11
in it and also includes updated cmake scripts that enable it by default.
But right now all the header files that exist in v18.18 still only require
C++03.  Let's see if anyone complains about the cmake changes or compiler
support for C++11 stuff in the DNN tooling.  If nothing bad happens then
I'll make C++11 required for the following dlib release and that release
can include the updated shape predictor.

On Fri, Jun 10, 2016 at 5:04 PM, Evgeniy Fominov <notifications@github.com>
wrote:

> Fixed compatibility for Visual Studio. Tested VS2010, VS2015, MINGW 5.3.
> Fixed performance leak in single thread mode (reason of problem -
> optimizer and CPU cache). Now trainer.set_num_threads(0 or 1) will make it
> work no slower than original implementation (really 20% faster)
>
> I am not finished checks with ibug - waiting for complete.
> 0 and 1 thread mode now have no difference in the testing scores with
> original version
> 2+ thread mode has some difference in the 4th digit normally. This is
> because rounding and limited float 32 precision. I can make it work with no
> difference, but performance affects very much
>
> My measurements on the examples\train_shape_predictor_ex (VS2015, x64)
> Threads Time, seconds Training error Testing error
> original 171 sec 2.40105e-05 0.085379
> new, 0-1 148 sec 2.40105e-05 0.085379
> 2 96 sec 0 0.0856668
> 3 70 sec 2.40105e-05 0.0856295
> 8 59 sec 2.40105e-05 0.0858176
>
> difference between 3 and 8 threads is small, because sample uses small
> amount of data. testing results on ibug dataset - tomorrow
> in 8 thread mode, CPU utilization on i7 is about 60-70%, because some
> parts are hard to make parallel
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/davisking/dlib/pull/126#issuecomment-225295415>, or mute
> the thread
> <https://github.com/notifications/unsubscribe/AF-Cx0vPxJ3WvInJIByScAnnYqvm035lks5qKdFygaJpZM4IsxMD>
> .
>