@e-fominov 2016-06-12
IBug results:
-----------------------------------------------------
|Method|Time|Training error|Testing error|
|---------|-------|-----------------|----------------|
|original|08h09m|0.0712444|0.0709394|
|new, 1 thread|7h 41 min|0.0712444|0.0709394|
|new, 4 threads|4h 18 min|0.071225|0.0708772|

Tested on Ubuntu + GCC 5.3. Memory usage is near identical
This code works too on VS 2010 and VS 2015

About C++11, DNN and Visual Studio - something is changing now in VS - new Uptade 3 for VS 2015 inproved  C++ support, but test_for_c++11 still does not compile
And NVidia published new Cuda 8 with support of VS 2015 Update 1 (Up2+ not supported)
VS 2015 compilation for DNN now has only a little amount of errors (only for layer_has_inpace... - like overloaded functions), so may be it will be possible to make VS 2015 compile code soon, after Cuda 8 will support VS 2015 Up3