@davisking 2016-06-12
Awesome, this is really nice work.  Did you try setting the block_size to a smaller value so that you get more than one block per thread?  I would expect that oversubscribing the thread pool to be faster, this way, if one of the threads finishes before the others it doesn't sit idle as long because there are probably other blocks to process.  So maybe having 8 times as many blocks as threads is often good.