@e-fominov 2016-06-14
Just checked with 8x blocks. Calculation time increased 2x. I am splitting very small parts of algorithm into threads, so there is too much overhead on thread management. And more than 4 threads will not give any profit. So 4 thread makes calculations 2x faster and this is the best I can make now. May be some situations with bigger trees or more features can have better results from more threds.

To get better results I need to rewrite full algorithm and this is not what we need now. Even 2x faster is a great profit