@davisking 2016-06-14
Ok, sounds good.  Yeah, 2x is definitely a good speedup.