@patricksnape 2016-07-21
@davisking Just a note about C++11 - if you make C++11 a hard requirement then you will lose Python 2.7 support as 2.7 still requires MSVC 2008.